Changed Decimal type convertion to BigDecimal in ClassGenerator

+++ datalake-spark31/src/main/scala/bio/ferlab/datalake/spark3/utils/ClassGenerator.scala
@@ -21,7 +21,7 @@ object ClassGenerator {
     case BooleanType                          => "Boolean"
     case DoubleType                           => "Double"
     case LongType                             => "Long"
-    case DecimalType()                        => "Double"
+    case DecimalType()                        => "BigDecimal"
     case DateType                             => "Date"
     case TimestampType                        => "Timestamp"
     case BinaryType                           => "Array[Byte]"
